requirements.txt
acres==0.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
dataclasses-json==0.6.7
distro==1.9.0
etelemetry==0.3.1
faiss-cpu==1.9.0.post1
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.1.0
frontend==0.0.3
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
groq==0.15.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-groq==0.2.3
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.5
langsmith==0.2.10
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.25.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.1
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
propcache==0.2.1
prov==2.0.1
puremagic==1.28
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydot==3.0.4
PyMuPDF==1.25.1
pyparsing==3.2.1
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytils==0.4.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
setuptools==75.8.0
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
tools==0.1.9
torch==2.5.1
tqdm==4.67.1
traits==6.4.3
transformers==4.48.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
